move stable map to utils
